--- internal/forge/github/connector.go
@@ -1,291 +1,0 @@
-package github
-
-import (
-	"context"
-	"errors"
-	"fmt"
-	"net/url"
-	"strconv"
-
-	"github.com/google/go-github/v58/github"
-	"golang.org/x/oauth2"
-
-	"github.com/git-town/git-town/v21/internal/browser"
-	"github.com/git-town/git-town/v21/internal/cli/print"
-	"github.com/git-town/git-town/v21/internal/forge/forgedomain"
-	"github.com/git-town/git-town/v21/internal/git/gitdomain"
-	"github.com/git-town/git-town/v21/internal/git/giturl"
-	"github.com/git-town/git-town/v21/internal/messages"
-	"github.com/git-town/git-town/v21/internal/subshell/subshelldomain"
-	"github.com/git-town/git-town/v21/pkg/colors"
-	. "github.com/git-town/git-town/v21/pkg/prelude"
-)
-
-// Connector provides standardized connectivity for the given repository (github.com/owner/repo)
-// via the GitHub API.
-type Connector struct {
-	forgedomain.Data
-	APIToken Option[forgedomain.GitHubToken]
-	client   *github.Client
-	log      print.Logger
-}
-
-func (self Connector) CreateProposal(data forgedomain.CreateProposalArgs) error {
-	browser.Open(self.NewProposalURL(data), data.FrontendRunner)
-	return nil
-}
-
-func (self Connector) DefaultProposalMessage(data forgedomain.ProposalData) string {
-	return DefaultProposalMessage(data)
-}
-
-func (self Connector) FindProposalFn() Option[func(branch, target gitdomain.LocalBranchName) (Option[forgedomain.Proposal], error)] {
-	if len(forgedomain.ReadProposalOverride()) > 0 {
-		return Some(self.findProposalViaOverride)
-	}
-	if self.APIToken.IsSome() {
-		return Some(self.findProposalViaAPI)
-	}
-	return None[func(branch, target gitdomain.LocalBranchName) (Option[forgedomain.Proposal], error)]()
-}
-
-func (self Connector) NewProposalURL(data forgedomain.CreateProposalArgs) string {
-	toCompare := data.Branch.String()
-	if data.ParentBranch != data.MainBranch {
-		toCompare = data.ParentBranch.String() + "..." + data.Branch.String()
-	}
-	result := fmt.Sprintf("%s/compare/%s?expand=1", self.RepositoryURL(), url.PathEscape(toCompare))
-	if title, hasTitle := data.ProposalTitle.Get(); hasTitle {
-		result += "&title=" + url.QueryEscape(title.String())
-	}
-	if body, hasBody := data.ProposalBody.Get(); hasBody {
-		result += "&body=" + url.QueryEscape(body.String())
-	}
-	return result
-}
-
-func (self Connector) OpenRepository(runner subshelldomain.Runner) error {
-	browser.Open(self.RepositoryURL(), runner)
-	return nil
-}
-
-func (self Connector) RepositoryURL() string {
-	return RepositoryURL(self.HostnameWithStandardPort(), self.Organization, self.Repository)
-}
-
-func (self Connector) SearchProposalFn() Option[func(gitdomain.LocalBranchName) (Option[forgedomain.Proposal], error)] {
-	if self.APIToken.IsNone() {
-		return None[func(gitdomain.LocalBranchName) (Option[forgedomain.Proposal], error)]()
-	}
-	return Some(self.searchProposal)
-}
-
-func (self Connector) SquashMergeProposalFn() Option[func(int, gitdomain.CommitMessage) error] {
-	if self.APIToken.IsNone() {
-		return None[func(int, gitdomain.CommitMessage) error]()
-	}
-	return Some(self.squashMergeProposal)
-}
-
-func (self Connector) UpdateProposalBodyFn() Option[func(forgedomain.ProposalInterface, string) error] {
-	if self.APIToken.IsNone() {
-		return None[func(forgedomain.ProposalInterface, string) error]()
-	}
-	return Some(self.updateProposalBody)
-}
-
-func (self Connector) UpdateProposalSourceFn() Option[func(forgedomain.ProposalInterface, gitdomain.LocalBranchName) error] {
-	return None[func(forgedomain.ProposalInterface, gitdomain.LocalBranchName) error]()
-}
-
-func (self Connector) UpdateProposalTargetFn() Option[func(forgedomain.ProposalInterface, gitdomain.LocalBranchName) error] {
-	if self.APIToken.IsNone() {
-		return None[func(forgedomain.ProposalInterface, gitdomain.LocalBranchName) error]()
-	}
-	return Some(self.updateProposalTarget)
-}
-
-func (self Connector) VerifyConnection() forgedomain.VerifyConnectionResult {
-	user, _, err := self.client.Users.Get(context.Background(), "")
-	if err != nil {
-		return forgedomain.VerifyConnectionResult{
-			AuthenticatedUser:   None[string](),
-			AuthenticationError: err,
-			AuthorizationError:  nil,
-		}
-	}
-	_, _, err = self.client.PullRequests.List(context.Background(), self.Organization, self.Repository, &github.PullRequestListOptions{
-		ListOptions: github.ListOptions{
-			PerPage: 1,
-		},
-	})
-	return forgedomain.VerifyConnectionResult{
-		AuthenticatedUser:   NewOption(*user.Login),
-		AuthenticationError: nil,
-		AuthorizationError:  err,
-	}
-}
-
-func (self Connector) findProposalViaAPI(branch, target gitdomain.LocalBranchName) (Option[forgedomain.Proposal], error) {
-	self.log.Start(messages.APIProposalLookupStart)
-	pullRequests, _, err := self.client.PullRequests.List(context.Background(), self.Organization, self.Repository, &github.PullRequestListOptions{
-		Head:  self.Organization + ":" + branch.String(),
-		Base:  target.String(),
-		State: "open",
-	})
-	if err != nil {
-		self.log.Failed(err.Error())
-		return None[forgedomain.Proposal](), err
-	}
-	if len(pullRequests) == 0 {
-		self.log.Success("none")
-		return None[forgedomain.Proposal](), nil
-	}
-	if len(pullRequests) > 1 {
-		return None[forgedomain.Proposal](), fmt.Errorf(messages.ProposalMultipleFromToFound, len(pullRequests), branch, target)
-	}
-	proposal := parsePullRequest(pullRequests[0])
-	self.log.Log(fmt.Sprintf("%s (%s)", colors.BoldGreen().Styled("#"+strconv.Itoa(proposal.Number)), proposal.Title))
-	return Some(forgedomain.Proposal{Data: proposal, ForgeType: forgedomain.ForgeTypeGitHub}), nil
-}
-
-func (self Connector) findProposalViaOverride(branch, target gitdomain.LocalBranchName) (Option[forgedomain.Proposal], error) {
-	self.log.Start(messages.APIProposalLookupStart)
-	proposalURLOverride := forgedomain.ReadProposalOverride()
-	self.log.Ok()
-	if proposalURLOverride == forgedomain.OverrideNoProposal {
-		return None[forgedomain.Proposal](), nil
-	}
-	return Some(forgedomain.Proposal{
-		Data: forgedomain.ProposalData{
-			Body:         None[string](),
-			MergeWithAPI: true,
-			Number:       123,
-			Source:       branch,
-			Target:       target,
-			Title:        "title",
-			URL:          proposalURLOverride,
-		},
-		ForgeType: forgedomain.ForgeTypeGitHub,
-	}), nil
-}
-
-func (self Connector) searchProposal(branch gitdomain.LocalBranchName) (Option[forgedomain.Proposal], error) {
-	self.log.Start(messages.APIParentBranchLookupStart, branch.String())
-	pullRequests, _, err := self.client.PullRequests.List(context.Background(), self.Organization, self.Repository, &github.PullRequestListOptions{
-		Head:  self.Organization + ":" + branch.String(),
-		State: "open",
-	})
-	if err != nil {
-		self.log.Failed(err.Error())
-		return None[forgedomain.Proposal](), err
-	}
-	if len(pullRequests) == 0 {
-		self.log.Success("none")
-		return None[forgedomain.Proposal](), nil
-	}
-	if len(pullRequests) > 1 {
-		return None[forgedomain.Proposal](), fmt.Errorf(messages.ProposalMultipleFromFound, len(pullRequests), branch)
-	}
-	proposal := parsePullRequest(pullRequests[0])
-	self.log.Success(proposal.Target.String())
-	return Some(forgedomain.Proposal{Data: proposal, ForgeType: forgedomain.ForgeTypeGitHub}), nil
-}
-
-func (self Connector) squashMergeProposal(number int, message gitdomain.CommitMessage) error {
-	if number <= 0 {
-		return errors.New(messages.ProposalNoNumberGiven)
-	}
-	self.log.Start(messages.ForgeGitHubMergingViaAPI, colors.BoldGreen().Styled("#"+strconv.Itoa(number)))
-	commitMessageParts := message.Parts()
-	_, _, err := self.client.PullRequests.Merge(context.Background(), self.Organization, self.Repository, number, commitMessageParts.Text, &github.PullRequestOptions{
-		MergeMethod: "squash",
-		CommitTitle: commitMessageParts.Subject,
-	})
-	if err != nil {
-		self.log.Ok()
-	}
-	return err
-}
-
-func (self Connector) updateProposalBody(proposalData forgedomain.ProposalInterface, updatedBody string) error {
-	data := proposalData.Data()
-	self.log.Start(messages.APIProposalUpdateBody, colors.BoldGreen().Styled("#"+strconv.Itoa(data.Number)))
-	_, _, err := self.client.PullRequests.Edit(context.Background(), self.Organization, self.Repository, data.Number, &github.PullRequest{
-		Body: Ptr(updatedBody),
-	})
-	if err != nil {
-		self.log.Failed(err.Error())
-		return err
-	}
-	self.log.Ok()
-	return nil
-}
-
-func (self Connector) updateProposalTarget(proposalData forgedomain.ProposalInterface, target gitdomain.LocalBranchName) error {
-	data := proposalData.Data()
-	targetName := target.String()
-	self.log.Start(messages.APIUpdateProposalTarget, colors.BoldGreen().Styled("#"+strconv.Itoa(data.Number)), colors.BoldCyan().Styled(targetName))
-	_, _, err := self.client.PullRequests.Edit(context.Background(), self.Organization, self.Repository, data.Number, &github.PullRequest{
-		Base: &github.PullRequestBranch{
-			Ref: &(targetName),
-		},
-	})
-	if err != nil {
-		self.log.Failed(err.Error())
-		return err
-	}
-	self.log.Ok()
-	return nil
-}
-
-func DefaultProposalMessage(data forgedomain.ProposalData) string {
-	return forgedomain.CommitBody(data, fmt.Sprintf("%s (#%d)", data.Title, data.Number))
-}
-
-type NewConnectorArgs struct {
-	APIToken  Option[forgedomain.GitHubToken]
-	Log       print.Logger
-	RemoteURL giturl.Parts
-}
-
-func NewConnector(args NewConnectorArgs) (Connector, error) {
-	tokenSource := oauth2.StaticTokenSource(&oauth2.Token{AccessToken: args.APIToken.String()})
-	httpClient := oauth2.NewClient(context.Background(), tokenSource)
-	githubClient := github.NewClient(httpClient)
-	if args.RemoteURL.Host != "github.com" {
-		url := "https://" + args.RemoteURL.Host
-		var err error
-		githubClient, err = githubClient.WithEnterpriseURLs(url, url)
-		if err != nil {
-			return Connector{}, fmt.Errorf(messages.GitHubEnterpriseInitializeError, err)
-		}
-	}
-	return Connector{
-		APIToken: args.APIToken,
-		Data: forgedomain.Data{
-			Hostname:     args.RemoteURL.Host,
-			Organization: args.RemoteURL.Org,
-			Repository:   args.RemoteURL.Repo,
-		},
-		client: githubClient,
-		log:    args.Log,
-	}, nil
-}
-
-func RepositoryURL(hostNameWithStandardPort string, organization string, repository string) string {
-	return fmt.Sprintf("https://%s/%s/%s", hostNameWithStandardPort, organization, repository)
-}
-
-// parsePullRequest extracts standardized proposal data from the given GitHub pull-request.
-func parsePullRequest(pullRequest *github.PullRequest) forgedomain.ProposalData {
-	return forgedomain.ProposalData{
-		Body:         NewOption(pullRequest.GetBody()),
-		Number:       pullRequest.GetNumber(),
-		Source:       gitdomain.NewLocalBranchName(pullRequest.Head.GetRef()),
-		Target:       gitdomain.NewLocalBranchName(pullRequest.Base.GetRef()),
-		Title:        pullRequest.GetTitle(),
-		MergeWithAPI: pullRequest.GetMergeableState() == "clean",
-		URL:          *pullRequest.HTMLURL,
-	}
-}
