+++ internal/forge/github/new_connector.go
@@ -1,0 +1,70 @@
+package github
+
+import (
+	"context"
+	"fmt"
+
+	"github.com/google/go-github/v58/github"
+	"golang.org/x/oauth2"
+
+	"github.com/git-town/git-town/v21/internal/cli/print"
+	"github.com/git-town/git-town/v21/internal/forge/forgedomain"
+	"github.com/git-town/git-town/v21/internal/git/gitdomain"
+	"github.com/git-town/git-town/v21/internal/git/giturl"
+	"github.com/git-town/git-town/v21/internal/messages"
+	. "github.com/git-town/git-town/v21/pkg/prelude"
+)
+
+type NewConnectorArgs struct {
+	APIToken  Option[forgedomain.GitHubToken]
+	Log       print.Logger
+	RemoteURL giturl.Parts
+}
+
+func NewConnector(args NewConnectorArgs) (forgedomain.Connector, error) {
+	anonConnector := AnonConnector{
+		Data: forgedomain.Data{
+			Hostname:     args.RemoteURL.Host,
+			Organization: args.RemoteURL.Org,
+			Repository:   args.RemoteURL.Repo,
+		},
+	}
+	apiToken, hasAPIToken := args.APIToken.Get()
+	if !hasAPIToken {
+		return anonConnector, nil
+	}
+	tokenSource := oauth2.StaticTokenSource(&oauth2.Token{AccessToken: apiToken.String()})
+	httpClient := oauth2.NewClient(context.Background(), tokenSource)
+	githubClient := github.NewClient(httpClient)
+	if args.RemoteURL.Host != "github.com" {
+		url := "https://" + args.RemoteURL.Host
+		var err error
+		githubClient, err = githubClient.WithEnterpriseURLs(url, url)
+		if err != nil {
+			return anonConnector, fmt.Errorf(messages.GitHubEnterpriseInitializeError, err)
+		}
+	}
+	return AuthConnector{
+		AnonConnector: anonConnector,
+		APIToken:      args.APIToken,
+		client:        githubClient,
+		log:           args.Log,
+	}, nil
+}
+
+func RepositoryURL(hostNameWithStandardPort string, organization string, repository string) string {
+	return fmt.Sprintf("https://%s/%s/%s", hostNameWithStandardPort, organization, repository)
+}
+
+// parsePullRequest extracts standardized proposal data from the given GitHub pull-request.
+func parsePullRequest(pullRequest *github.PullRequest) forgedomain.ProposalData {
+	return forgedomain.ProposalData{
+		Body:         NewOption(pullRequest.GetBody()),
+		Number:       pullRequest.GetNumber(),
+		Source:       gitdomain.NewLocalBranchName(pullRequest.Head.GetRef()),
+		Target:       gitdomain.NewLocalBranchName(pullRequest.Base.GetRef()),
+		Title:        pullRequest.GetTitle(),
+		MergeWithAPI: pullRequest.GetMergeableState() == "clean",
+		URL:          *pullRequest.HTMLURL,
+	}
+}
