+++ internal/forge/github/proposal_finder.go
@@ -1,0 +1,67 @@
+package github
+
+import (
+	"context"
+	"fmt"
+	"strconv"
+
+	"github.com/git-town/git-town/v21/internal/forge/forgedomain"
+	"github.com/git-town/git-town/v21/internal/git/gitdomain"
+	"github.com/git-town/git-town/v21/internal/messages"
+	"github.com/git-town/git-town/v21/pkg/colors"
+	. "github.com/git-town/git-town/v21/pkg/prelude"
+	"github.com/google/go-github/v58/github"
+)
+
+// TODO: make sure all AuthConnector instances use the pointer receiver, so that we don't copy the connector instance.
+// Or use a Mutable for it.
+func (self *AuthConnector) FindProposal(branch, target gitdomain.LocalBranchName) (Option[forgedomain.Proposal], error) {
+	if len(forgedomain.ReadProposalOverride()) > 0 {
+		return self.findProposalViaOverride(branch, target)
+	}
+	return self.findProposalViaAPI(branch, target)
+}
+
+func (self AuthConnector) findProposalViaAPI(branch, target gitdomain.LocalBranchName) (Option[forgedomain.Proposal], error) {
+	self.log.Start(messages.APIProposalLookupStart)
+	pullRequests, _, err := self.client.PullRequests.List(context.Background(), self.Organization, self.Repository, &github.PullRequestListOptions{
+		Head:  self.Organization + ":" + branch.String(),
+		Base:  target.String(),
+		State: "open",
+	})
+	if err != nil {
+		self.log.Failed(err.Error())
+		return None[forgedomain.Proposal](), err
+	}
+	if len(pullRequests) == 0 {
+		self.log.Success("none")
+		return None[forgedomain.Proposal](), nil
+	}
+	if len(pullRequests) > 1 {
+		return None[forgedomain.Proposal](), fmt.Errorf(messages.ProposalMultipleFromToFound, len(pullRequests), branch, target)
+	}
+	proposal := parsePullRequest(pullRequests[0])
+	self.log.Log(fmt.Sprintf("%s (%s)", colors.BoldGreen().Styled("#"+strconv.Itoa(proposal.Number)), proposal.Title))
+	return Some(forgedomain.Proposal{Data: proposal, ForgeType: forgedomain.ForgeTypeGitHub}), nil
+}
+
+func (self AuthConnector) findProposalViaOverride(branch, target gitdomain.LocalBranchName) (Option[forgedomain.Proposal], error) {
+	self.log.Start(messages.APIProposalLookupStart)
+	proposalURLOverride := forgedomain.ReadProposalOverride()
+	self.log.Ok()
+	if proposalURLOverride == forgedomain.OverrideNoProposal {
+		return None[forgedomain.Proposal](), nil
+	}
+	return Some(forgedomain.Proposal{
+		Data: forgedomain.ProposalData{
+			Body:         None[string](),
+			MergeWithAPI: true,
+			Number:       123,
+			Source:       branch,
+			Target:       target,
+			Title:        "title",
+			URL:          proposalURLOverride,
+		},
+		ForgeType: forgedomain.ForgeTypeGitHub,
+	}), nil
+}
