+++ internal/forge/github/proposal_merger.go
@@ -1,0 +1,28 @@
+package github
+
+import (
+	"context"
+	"errors"
+	"strconv"
+
+	"github.com/git-town/git-town/v21/internal/git/gitdomain"
+	"github.com/git-town/git-town/v21/internal/messages"
+	"github.com/git-town/git-town/v21/pkg/colors"
+	"github.com/google/go-github/v58/github"
+)
+
+func (self AuthConnector) SquashMergeProposal(number int, message gitdomain.CommitMessage) error {
+	if number <= 0 {
+		return errors.New(messages.ProposalNoNumberGiven)
+	}
+	self.log.Start(messages.ForgeGitHubMergingViaAPI, colors.BoldGreen().Styled("#"+strconv.Itoa(number)))
+	commitMessageParts := message.Parts()
+	_, _, err := self.client.PullRequests.Merge(context.Background(), self.Organization, self.Repository, number, commitMessageParts.Text, &github.PullRequestOptions{
+		MergeMethod: "squash",
+		CommitTitle: commitMessageParts.Subject,
+	})
+	if err != nil {
+		self.log.Ok()
+	}
+	return err
+}
