+++ internal/forge/github/proposal_searcher.go
@@ -1,0 +1,34 @@
+package github
+
+import (
+	"context"
+	"fmt"
+
+	"github.com/git-town/git-town/v21/internal/forge/forgedomain"
+	"github.com/git-town/git-town/v21/internal/git/gitdomain"
+	"github.com/git-town/git-town/v21/internal/messages"
+	. "github.com/git-town/git-town/v21/pkg/prelude"
+	"github.com/google/go-github/v58/github"
+)
+
+func (self AuthConnector) SearchProposal(branch gitdomain.LocalBranchName) (Option[forgedomain.Proposal], error) {
+	self.log.Start(messages.APIParentBranchLookupStart, branch.String())
+	pullRequests, _, err := self.client.PullRequests.List(context.Background(), self.Organization, self.Repository, &github.PullRequestListOptions{
+		Head:  self.Organization + ":" + branch.String(),
+		State: "open",
+	})
+	if err != nil {
+		self.log.Failed(err.Error())
+		return None[forgedomain.Proposal](), err
+	}
+	if len(pullRequests) == 0 {
+		self.log.Success("none")
+		return None[forgedomain.Proposal](), nil
+	}
+	if len(pullRequests) > 1 {
+		return None[forgedomain.Proposal](), fmt.Errorf(messages.ProposalMultipleFromFound, len(pullRequests), branch)
+	}
+	proposal := parsePullRequest(pullRequests[0])
+	self.log.Success(proposal.Target.String())
+	return Some(forgedomain.Proposal{Data: proposal, ForgeType: forgedomain.ForgeTypeGitHub}), nil
+}
